+++ src/bridge-canister/src/runtime/scheduler.rs
@@ -7,7 +7,6 @@ use bridge_did::event_data::*;
 use bridge_did::op_id::OperationId;
 use bridge_utils::bft_events::BridgeEvent;
 use candid::{CandidType, Decode};
-use drop_guard::guard;
 use ic_stable_structures::stable_structures::Memory;
 use ic_stable_structures::{StableBTreeMap, StableCell};
 use ic_task_scheduler::scheduler::{Scheduler, TaskScheduler};
@@ -100,12 +99,9 @@ pub fn log_task_execution_error<Op: Operation>(task: InnerScheduledTask<BridgeTa
 
 /// Task type used by `BridgeRuntime`.
 #[derive(Debug, Clone, Serialize, Deserialize, CandidType)]
-pub enum BridgeTask<Op> {
-    /// Bridge operations defined by user.
-    Operation(OperationId, Op),
-
-    /// Bridge operations defined by the runtime itself.
-    Service(ServiceTask),
+pub struct BridgeTask<Op> {
+    op_id: OperationId,
+    operation: Op,
 }
 
 impl<Op: Operation> BridgeTask<Op> {
@@ -114,206 +110,48 @@ impl<Op: Operation> BridgeTask<Op> {
         ctx: RuntimeState<Op>,
         task_scheduler: DynScheduler<Op>,
     ) -> BftResult<()> {
-        match self {
-            BridgeTask::Operation(op_id, _) => {
-                let Some(operation) = ctx.borrow().operations.get(op_id) else {
-                    log::warn!("Operation #{op_id} not found.");
-                    return Err(Error::OperationNotFound(op_id));
-                };
+        let Some(operation) = ctx.borrow().operations.get(self.op_id) else {
+            log::warn!("Operation #{} not found.", { self.op_id });
+            return Err(Error::OperationNotFound(self.op_id));
+        };
 
-                let ctx_clone = ctx.clone();
-                let progress =
-                    operation
-                        .progress(op_id, ctx.clone())
-                        .await
-                        .inspect_err(move |err| {
-                            ctx_clone
-                                .borrow_mut()
-                                .operations
-                                .update_with_err(op_id, err.to_string())
-                        })?;
+        let ctx_clone = ctx.clone();
+        let progress = operation
+            .progress(self.op_id, ctx.clone())
+            .await
+            .inspect_err(move |err| {
+                ctx_clone
+                    .borrow_mut()
+                    .operations
+                    .update_with_err(self.op_id, err.to_string())
+            })?;
 
-                let new_op = match progress {
-                    OperationProgress::Progress(op) => op,
-                    OperationProgress::AddToService(service_id) => {
-                        ctx.borrow().push_operation_to_service(service_id, op_id)?;
-                        return Ok(());
-                    }
-                };
-
-                let scheduling_options = new_op.scheduling_options();
-                ctx.borrow_mut().operations.update(op_id, new_op.clone());
-
-                if let Some(options) = scheduling_options {
-                    let scheduled_task =
-                        ScheduledTask::with_options(Self::Operation(op_id, new_op), options);
-                    task_scheduler.append_task(scheduled_task);
-                }
-
-                Ok(())
+        let new_op = match progress {
+            OperationProgress::Progress(op) => op,
+            OperationProgress::AddToService(service_id) => {
+                ctx.borrow()
+                    .push_operation_to_service(service_id, self.op_id)?;
+                return Ok(());
             }
-            BridgeTask::Service(service_task) => service_task.execute(ctx, task_scheduler).await,
-        }
-    }
-}
+        };
 
-/// Service tasks, done by the `BridgeRuntime` by default.
-#[derive(Debug, Clone, Serialize, Deserialize, CandidType)]
-pub enum ServiceTask {
-    /// Task to query logs from EVM.
-    CollectEvmLogs,
+        let scheduling_options = new_op.scheduling_options();
+        ctx.borrow_mut()
+            .operations
+            .update(self.op_id, new_op.clone());
 
-    /// Task to refresh EVM parameters.
-    RefreshEvmParams,
-}
-
-impl ServiceTask {
-    const MAX_LOG_REQUEST_COUNT: u64 = 1000;
-
-    async fn execute<Op: Operation>(
-        self,
-        ctx: RuntimeState<Op>,
-        task_scheduler: DynScheduler<Op>,
-    ) -> BftResult<()> {
-        match self {
-            ServiceTask::CollectEvmLogs => {
-                let _lock = guard(ctx.clone(), |s| s.borrow_mut().collecting_logs_ts = None);
-
-                ServiceTask::collect_evm_logs(ctx.clone(), task_scheduler).await
-            }
-            ServiceTask::RefreshEvmParams => {
-                let _lock = guard(ctx.clone(), |s| {
-                    s.borrow_mut().refreshing_evm_params_ts = None
-                });
-                let config = ctx.borrow().config.clone();
-                ConfigStorage::refresh_evm_params(config).await
-            }
-        }
-    }
-
-    async fn collect_evm_logs<Op: Operation>(
-        ctx: RuntimeState<Op>,
-        task_scheduler: DynScheduler<Op>,
-    ) -> BftResult<()> {
-        let collected = ctx.collect_evm_events(Self::MAX_LOG_REQUEST_COUNT).await?;
-        let events = collected.events;
-
-        ctx.borrow()
-            .config
-            .borrow_mut()
-            .update_evm_params(|params| params.next_block = collected.last_block_number + 1);
-
-        for event in events {
-            let operation_action = match event {
-                BridgeEvent::Burnt(event) => Op::on_wrapped_token_burnt(ctx.clone(), event).await,
-                BridgeEvent::Minted(event) => Op::on_wrapped_token_minted(ctx.clone(), event).await,
-                BridgeEvent::Notify(event) => {
-                    Self::on_minter_notification(ctx.clone(), event, &task_scheduler).await
-                }
-            };
-
-            let to_schedule = match operation_action {
-                Some(OperationAction::Create(op, memo)) => {
-                    let new_op_id = ctx.borrow_mut().operations.new_operation(op.clone(), memo);
-                    op.scheduling_options().zip(Some((new_op_id, op)))
-                }
-                Some(OperationAction::CreateWithId(id, op, memo)) => {
-                    ctx.borrow_mut()
-                        .operations
-                        .new_operation_with_id(id, op.clone(), memo);
-                    op.scheduling_options().zip(Some((id, op)))
-                }
-                Some(OperationAction::Update { nonce, update_to }) => {
-                    let Some((operation_id, _)) = ctx
-                        .borrow()
-                        .operations
-                        .get_for_address(&update_to.evm_wallet_address(), None)
-                        .into_iter()
-                        .find(|(operation_id, _)| operation_id.nonce() == nonce)
-                    else {
-                        log::warn!(
-                            "operation with dst_address = {} and nonce {} not found",
-                            update_to.evm_wallet_address(),
-                            nonce
-                        );
-                        return Err(Error::OperationNotFound(OperationId::new(nonce as _)));
-                    };
-
-                    ctx.borrow_mut()
-                        .operations
-                        .update(operation_id, update_to.clone());
-                    update_to
-                        .scheduling_options()
-                        .zip(Some((operation_id, update_to)))
-                }
-                None => None,
-            };
-
-            if let Some((options, (op_id, op))) = to_schedule {
-                let task = ScheduledTask::with_options(BridgeTask::Operation(op_id, op), options);
-                task_scheduler.append_task(task);
-            }
-        }
-
-        log::debug!("EVM logs collected");
-        Ok(())
-    }
-
-    async fn on_minter_notification<Op: Operation>(
-        ctx: RuntimeState<Op>,
-        data: NotifyMinterEventData,
-        scheduler: &DynScheduler<Op>,
-    ) -> Option<OperationAction<Op>> {
-        match data.notification_type {
-            MinterNotificationType::RescheduleOperation => {
-                let operation_id = match Decode!(&data.user_data, OperationId) {
-                    Ok(v) => v,
-                    Err(err) => {
-                        log::warn!("Failed to decode operation id from reschedule operation request: {err:?}");
-                        return None;
-                    }
-                };
-
-                Self::reschedule_operation(ctx, operation_id, scheduler);
-                None
-            }
-            _ => Op::on_minter_notification(ctx, data).await,
-        }
-    }
-
-    fn reschedule_operation<Op: Operation>(
-        ctx: RuntimeState<Op>,
-        operation_id: OperationId,
-        scheduler: &DynScheduler<Op>,
-    ) {
-        let Some(operation) = ctx.borrow().operations.get(operation_id) else {
-            log::warn!(
-                "Reschedule of operation #{operation_id} is requested but it does not exist"
+        if let Some(options) = scheduling_options {
+            let scheduled_task = ScheduledTask::with_options(
+                Self {
+                    op_id: self.op_id,
+                    operation: new_op,
+                },
+                options,
             );
-            return;
-        };
-
-        let Some(task_options) = operation.scheduling_options() else {
-            log::info!("Reschedule of operation #{operation_id} is requested but no scheduling is required for this operation");
-            return;
-        };
-
-        let current_task_id = scheduler.find_id(&|op| match op {
-            BridgeTask::Operation(id, _) => id == operation_id,
-            BridgeTask::Service(_) => false,
-        });
-        match current_task_id {
-            Some(task_id) => {
-                scheduler.reschedule(task_id, task_options.clone());
-                log::trace!("Updated schedule for operation #{operation_id} task #{task_id} to {task_options:?}");
-            }
-            None => {
-                let task_id = scheduler.append_task(
-                    (BridgeTask::Operation(operation_id, operation), task_options).into(),
-                );
-                log::trace!("Restarted operation #{operation_id} with task id #{task_id}");
-            }
+            task_scheduler.append_task(scheduled_task);
         }
+
+        Ok(())
     }
 }
 
@@ -447,7 +285,10 @@ mod tests {
         const COUNT: usize = 5;
         for _ in 0..COUNT {
             let op = ctx.borrow().operations.get(id).unwrap();
-            let task = BridgeTask::Operation(id, op);
+            let task = BridgeTask {
+                op_id: id,
+                operation: op,
+            };
             task.execute_inner(ctx.clone(), Box::new(runtime.scheduler.clone()))
                 .await
                 .unwrap_err();
@@ -482,7 +323,10 @@ mod tests {
         const COUNT: usize = 5;
         for _ in 0..COUNT {
             let op = ctx.borrow().operations.get(id).unwrap();
-            let task = BridgeTask::Operation(id, op);
+            let task = BridgeTask {
+                op_id: id,
+                operation: op,
+            };
             task.execute_inner(ctx.clone(), Box::new(runtime.scheduler.clone()))
                 .await
                 .unwrap();
@@ -517,7 +361,10 @@ mod tests {
         let op = TestOperation::new_err();
         let id = ctx.borrow_mut().operations.new_operation(op.clone(), None);
 
-        let task = BridgeTask::Operation(id, op);
+        let task = BridgeTask {
+            op_id: id,
+            operation: op,
+        };
         let err = task
             .execute(ctx.clone(), Box::new(runtime.scheduler.clone()))
             .await
